# contract/openapi.yaml
openapi: 3.1.0
info:
  title: GoalFather API
  version: 0.1.0
  description: |
    Contrato compartilhado entre o backend (Kotlin + Spring Boot) e o frontend
    (Vite + React + TypeScript). Esta é a fonte de verdade dos tipos da API.

    - **Backend:** springdoc-openapi valida que a implementação Spring bate com
      este spec. `sealed interface MatchEvent` em Kotlin é serializado com
      `@JsonClassDiscriminator("type")` (kotlinx.serialization), mapeando
      diretamente para `oneOf + discriminator` aqui.
    - **Frontend:** `openapi-typescript` gera `src/api/generated.d.ts` a partir
      deste arquivo. Mocks MSW e cliente HTTP usam esses tipos.

    Mudanças no contrato são feitas aqui PRIMEIRO; backend e frontend ajustam
    suas implementações para acompanhar.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Fabio Carlesso
    email: fabio.carlesso@gmail.com
servers:
  - url: http://localhost:8080
    description: Backend Spring Boot local (bootRun)
  - url: http://localhost:5173
    description: Frontend Vite com MSW interceptando (dev mock)
...
paths:
...
  /api/clubs/available:
    get:
      tags: [Club]
      operationId: listAvailableClubs
      summary: Clubes sem dono (disponíveis para reivindicar)
      description: |
        Lista os clubes com `ownerId == null`, com metadados para o card de
        seleção (caixa, capacidade do estádio, força média do elenco).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Clubes disponíveis
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AvailableClub' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/clubs/{id}/claim:
    post:
      tags: [Club]
      operationId: claimClub
      summary: Reivindicar um clube para o usuário autenticado
      description: |
        Marca o clube como pertencente ao usuário autenticado e grava o
        `clubId` no usuário. Transacional com lock otimista: se dois usuários
        tentarem reivindicar o mesmo clube simultaneamente, o segundo recebe
        `409 CLUB_ALREADY_CLAIMED`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClubIdPath'
      responses:
        '200':
          description: Clube reivindicado — devolve o usuário atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthUser' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Clube já reivindicado (`CLUB_ALREADY_CLAIMED`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/clubs/{id}:
    get:
      tags: [Club]
      operationId: getClub
      summary: Estado completo do clube
      description: Retorna clube com elenco, caixa, estádio e escalação atual.
      parameters:
        - $ref: '#/components/parameters/ClubIdPath'
      responses:
        '200':
          description: Clube encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Club' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/clubs/{id}/lineup:
    post:
      tags: [Club]
      operationId: saveLineup
      summary: Salvar escalação, formação e postura
      description: |
        Define os 11 titulares, a formação e a postura da partida. A ordem do
        array `playerIds` deve corresponder aos slots da formação escolhida
        (ver `Formation`).

        A postura (issue #56) é gravada JUNTO da escalação, numa única
        operação — a engine da rodada lê as duas do mesmo estado persistido.
        Omitir `posture` mantém `BALANCED`.
      parameters:
        - $ref: '#/components/parameters/ClubIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LineupRequest' }
      responses:
        '204':
          description: Escalação salva
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/clubs/{id}/stadium/expand:
    post:
      tags: [Stadium]
      operationId: expandStadium
      summary: Ampliar capacidade do estádio
      description: |
        Investe caixa do clube para aumentar a capacidade do estádio.
        Falha com `INSUFFICIENT_FUNDS` se o caixa for insuficiente.
      parameters:
        - $ref: '#/components/parameters/ClubIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExpandStadiumRequest' }
      responses:
        '200':
          description: Estádio ampliado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Club' }
        '402':
          description: Caixa insuficiente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/clubs/{id}/medical:
    post:
      tags: [Medical]
      operationId: treatSquad
      summary: Sessão do departamento médico
      description: |
        Paga a taxa fixa de R$ 30.000 (3000000 centavos) e acelera a
        recuperação do elenco (issue #54): todo jogador ganha +30 de stamina
        (teto 100) e as lesões em curso encurtam uma rodada. Falha com
        `INSUFFICIENT_FUNDS` se o caixa não cobrir a taxa.
      parameters:
        - $ref: '#/components/parameters/ClubIdPath'
      responses:
        '200':
          description: Elenco tratado — devolve o clube já atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Club' }
        '402':
          description: Caixa insuficiente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Usuário não é dono do clube
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/clubs/{id}/training:
    post:
      tags: [Training]
      operationId: setTrainingFocus
      summary: Escolher o foco de treino da semana
      description: |
        Registra o foco de treino do clube (issue #58). O efeito não acontece
        aqui: ele entra na virada da rodada, junto do desgaste físico — e
        exatamente uma vez, sob o mesmo claim que protege os demais efeitos.
        Quem nunca escolheu (e todo clube da IA) treina em `DESCANSO`.
      parameters:
        - $ref: '#/components/parameters/ClubIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TrainingFocusRequest' }
      responses:
        '200':
          description: Foco registrado — devolve o clube já atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Club' }
        '403':
          description: Usuário não é dono do clube
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/clubs/{id}/ticket-price:
    put:
      tags: [Stadium]
      operationId: setTicketPrice
      summary: Definir o preço do ingresso do estádio
      description: |
        Define quanto o clube cobra por ingresso (issue #59). `PUT`, e não
        `POST` como os comandos vizinhos, porque é a substituição idempotente
        de um ajuste do clube, não uma ação com efeito acumulado.

        O preço alimenta a curva de demanda da bilheteria: acima do "preço
        justo" do time o estádio esvazia progressivamente, abaixo dele a
        ocupação se aproxima do teto de 100%. Como a ocupação satura, a receita
        tem um máximo em preço intermediário — e o preço justo sobe com a força
        do elenco, então time melhor suporta ingresso mais caro.

        Fora da faixa `10_00..200_00` (centavos) o comando falha com
        `TICKET_PRICE_OUT_OF_RANGE`. Quem nunca mexeu — e todo clube da IA —
        cobra R$ 50 (`5000`).
      parameters:
        - $ref: '#/components/parameters/ClubIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TicketPriceRequest' }
      responses:
        '200':
          description: Preço registrado — devolve o clube já atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Club' }
        '400':
          description: Preço fora da faixa permitida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Usuário não é dono do clube
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT emitido por `/api/auth/register` ou `/api/auth/login`. Enviado em
        `Authorization: Bearer <token>`. Todos os `/api/**` exigem o token,
        exceto `/api/auth/register` e `/api/auth/login`.
  parameters:
    ClubIdPath:
      name: id
      in: path
      required: true
      description: ID do clube
      schema: { type: integer, format: int64 }
  responses:
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Payload inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Token ausente ou inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    AuthUser:
      type: object
      required: [id, username]
      properties:
        id:       { type: integer, format: int64 }
        username: { type: string }
        clubId:
          type: integer
          format: int64
          nullable: true
          description: Clube reivindicado pelo usuário; `null` antes da seleção.
...
    AvailableClub:
      description: Resumo de um clube sem dono, para o card de seleção.
      type: object
      required: [id, name, cash, stadiumCapacity, averageOverall, division]
      properties:
        id:              { type: integer, format: int64 }
        name:            { type: string }
        cash:            { type: integer, format: int64, description: Caixa em centavos }
        stadiumCapacity: { type: integer, minimum: 0 }
        averageOverall:
          type: integer
          minimum: 0
          maximum: 99
          description: Força média do elenco (média dos `overall`).
        division:
          type: integer
          minimum: 1
          description: 'Divisão em que o clube disputa a temporada (1 = elite)'

    # ─── Enums ─────────────────────────────────────────────────────────────
    Position:
      type: string
      description: Posição do jogador no campo
      enum: [GK, CB, MF, FW]
      x-labels:
        GK: GL
        CB: ZG
        MF: MC
        FW: AT
    Formation:
      type: string
      description: |
        Esquema tático (define ordem das posições nos slots). Desde a issue #56
        a formação também pesa na simulação: 4-3-3 e 3-5-2 inclinam o time ao
        ataque, 5-3-2 à defesa e 4-4-2 é o eixo neutro. O peso é menor que o da
        `Posture` — a formação inclina, a postura decide.
      enum: ['4-4-2', '4-3-3', '3-5-2', '5-3-2']
    Posture:
      type: string
      description: |
        Postura do time na partida (issue #56) — a decisão tática por rodada,
        salva junto da escalação.

        `OFFENSIVE` aumenta a chance de gol dos DOIS times (jogo aberto);
        `DEFENSIVE` reduz as duas, cedendo menos do que deixa de criar — é o
        que torna "fechar o time" racional contra adversário mais forte.
        Os multiplicadores vivem no backend (`enum class Posture`); o cliente
        só escolhe e exibe.
      enum: [DEFENSIVE, BALANCED, OFFENSIVE]
      default: BALANCED
      x-labels:
        DEFENSIVE: Defensiva
        BALANCED: Equilibrada
        OFFENSIVE: Ofensiva
    TrainingFocus:
      type: string
      description: |
        Foco do treino da semana (issue #58) — a decisão de gestão ENTRE
        rodadas, diferente da `Posture`, que é a decisão da partida.

        `ATAQUE`/`DEFESA` dão a alguns jogadores uma chance pequena de +1 no
        atributo correspondente (jovens aproveitam mais), devolvem pouca forma
        física e cobram risco pequeno de lesão leve. `FISICO` fica no meio:
        recupera bem a stamina, desenvolve velocidade e é o que mais machuca.
        `DESCANSO` é o único sem risco, e o que mais recupera. Nenhum foco
        apaga o desgaste de quem jogou (a fadiga continua acumulando), e
        nenhum recupera zero — foco técnico custa forma física, mas não deixa
        o time preso no piso. Os números vivem no backend
        (`enum class TrainingFocus`); o cliente só escolhe e exibe.
      enum: [ATAQUE, DEFESA, FISICO, DESCANSO]
      default: DESCANSO
      x-labels:
        ATAQUE: Ataque
        DEFESA: Defesa
        FISICO: Físico
        DESCANSO: Descanso
...
    Player:
      type: object
      required:
        - id
        - name
        - position
        - overall
        - pace
        - shooting
        - passing
        - defending
        - stamina
        - salary
        - age
        - goals
        - yellowCards
        - redCards
        - availability
      properties:
        id:         { type: integer, format: int64 }
        name:       { type: string }
        position:   { $ref: '#/components/schemas/Position' }
        overall:    { type: integer, minimum: 0, maximum: 99 }
        pace:       { type: integer, minimum: 0, maximum: 99 }
        shooting:   { type: integer, minimum: 0, maximum: 99 }
        passing:    { type: integer, minimum: 0, maximum: 99 }
        defending:  { type: integer, minimum: 0, maximum: 99 }
        stamina:
          type: integer
          minimum: 0
          maximum: 100
          description: |
            Forma física (issue #54). Titulares perdem 10–25 por rodada (piso
            40) e reservas recuperam. Abaixo de 70 o rendimento do jogador cai
            proporcionalmente — a UI não recalcula isso, só exibe.
        salary:     { type: integer, minimum: 0, description: Salário mensal }
        age:
          type: integer
          minimum: 15
          maximum: 50
          description: |
            Idade (issue #55). Na virada de temporada todo jogador ganha um ano:
            até 23 tende a evoluir, de 24 a 29 oscila pouco e a partir dos 30
            tende a regredir. Veterano de overall baixo se aposenta e sai do
            elenco — a UI não recalcula nada disso, só exibe.
        goals:      { type: integer, minimum: 0, description: Gols na temporada }
        yellowCards: { type: integer, minimum: 0, description: Cartões amarelos na temporada }
        redCards:    { type: integer, minimum: 0, description: Cartões vermelhos na temporada }
        availability: { $ref: '#/components/schemas/Availability' }
        star:
          type: boolean
          description: Derivado de `overall >= 82` (computado pelo backend)
    Availability:
      description: |
        Disponibilidade do jogador (issue #54). `oneOf` + `discriminator`
        espelhando o `sealed interface Availability` do domínio Kotlin: a
        duração da lesão só existe DENTRO da variante `Injured`, então não há
        como representar "lesionado sem duração". No TS, desambigua com
        `switch (player.availability.type)` exaustivo.
      oneOf:
        - $ref: '#/components/schemas/AvailableStatus'
        - $ref: '#/components/schemas/InjuredStatus'
      discriminator:
        propertyName: type
        mapping:
          Available: '#/components/schemas/AvailableStatus'
          Injured:   '#/components/schemas/InjuredStatus'
    AvailableStatus:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [Available] }
    InjuredStatus:
      type: object
      required: [type, roundsOut]
      properties:
        type: { type: string, enum: [Injured] }
        roundsOut:
          type: integer
          minimum: 1
          description: Rodadas restantes de afastamento; chega a zero e o jogador volta.
    Lineup:
      type: object
      required: [formation, playerIds]
      properties:
        formation:
          $ref: '#/components/schemas/Formation'
        playerIds:
          type: array
          minItems: 0
          maxItems: 11
          items: { type: integer, format: int64 }
          description: |
            IDs dos titulares em ordem correspondente aos slots da formação.
        posture:
          $ref: '#/components/schemas/Posture'
          description: |
            Postura salva com a escalação (issue #56). O backend sempre a
            devolve — escalação gravada antes da tática existir desserializa
            no default. Opcional no schema porque o cliente pode ter omitido
            ao salvar; ausente, trate como `BALANCED`.
    Club:
      type: object
      required:
        - id
        - name
        - cash
        - stadiumCapacity
        - squad
        # Sempre presente: quem nunca escolheu vem com o default `DESCANSO`,
        # então o cliente nunca precisa adivinhar o foco em vigor.
        - trainingFocus
        # Idem: quem nunca mexeu no ingresso vem com R$ 50 (`5000`).
        - ticketPriceCents
      properties:
        id:               { type: integer, format: int64 }
        name:             { type: string }
        cash:             { type: integer, format: int64, description: Caixa em centavos }
        stadiumCapacity:  { type: integer, minimum: 0 }
        squad:
          type: array
          items: { $ref: '#/components/schemas/Player' }
        lineup:
          $ref: '#/components/schemas/Lineup'
          description: Escalação salva (pode estar incompleta)
        ownerId:
          type: integer
          format: int64
          nullable: true
          description: |
            Dono humano do clube; `null` significa controlado pela IA.
            Preparado para multiplayer (Fase 5).
        trainingFocus:
          $ref: '#/components/schemas/TrainingFocus'
          description: |
            Foco de treino em vigor (issue #58). Vale entre rodadas e para o
            elenco inteiro — fica valendo até o técnico trocar.
        ticketPriceCents:
          type: integer
          format: int64
          minimum: 1000
          maximum: 20000
          description: |
            Preço do ingresso em vigor, em centavos (issue #59). Vale até o
            técnico trocar; quem nunca mexeu cobra `5000` (R$ 50).

    # ─── Requests ─────────────────────────────────────────────────────────
    LineupRequest:
      type: object
      required: [formation, playerIds]
      properties:
        formation: { $ref: '#/components/schemas/Formation' }
        playerIds:
          type: array
          minItems: 11
          maxItems: 11
          items: { type: integer, format: int64 }
        posture:
          $ref: '#/components/schemas/Posture'
          description: Postura da partida; omitida, o backend assume `BALANCED`.
    TrainingFocusRequest:
      type: object
      required: [focus]
      properties:
        focus: { $ref: '#/components/schemas/TrainingFocus' }
    TicketPriceRequest:
      type: object
      required: [ticketPriceCents]
      properties:
        ticketPriceCents:
          type: integer
          format: int64
          minimum: 1000
          maximum: 20000
          description: |
            Novo preço do ingresso em centavos (issue #59). Fora de
            `1000..20000` o backend responde 400 `TICKET_PRICE_OUT_OF_RANGE` —
            a faixa é regra de domínio, o `minimum`/`maximum` aqui só documenta
            e ajuda o cliente a validar antes de mandar.
    ExpandStadiumRequest:
      type: object
      required: [additionalSeats]
      properties:
        additionalSeats:
          type: integer
          minimum: 1000
          description: Quantidade de novos assentos a adicionar
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Código estável para o cliente discriminar (não a mensagem).
            Ex.: `CLUB_NOT_FOUND`, `INVALID_LINEUP`, `INSUFFICIENT_FUNDS`.
        message:
          type: string
          description: Mensagem humana (pode ser exibida ao usuário)
        details:
          type: object
          additionalProperties: true
          description: Dados extras dependentes do código de erro